requirements.txt
pycryptodome
sympy

